moved run and register functions out, swapped to cli-interfaces
