+++ package.json
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "codecov.io": "0.1.6",
+    "dojo-cli": "^2.0.0-alpha.3",
     "dojo-loader": ">=2.0.0-beta",
     "dts-generator": "~1.7.0",
     "glob": "^7.0.3",
