+++ src/changeDir.ts
@@ -1,0 +1,3 @@
+export default function(path: string) {
+	process.chdir(path);
+}
