+++ src/dirname.ts
@@ -1,0 +1,1 @@
+export default __dirname;
